pyproject.toml
[project]
name = "causal-ssm-agent"
version = "0.1.0"
description = "LLM-orchestrated causal inference over long-context, multi-source data"
readme = "README.md"
authors = [
    { name = "ma9o", email = "magogagiovanni@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "arviz>=0.19.0",
    "blackjax>=1.2.0",
    "cuthbert>=0.0.3",
    "equinox>=0.11.0",
    "exa-py>=1.0.0",
    "google-genai>=1.56.0",
    "inspect-ai>=0.3.162",
    "jax>=0.4.35",
    "jaxlib>=0.4.35",
    "networkx>=3.6",
    "numpyro>=0.16.0",
    "optax>=0.2.0",
    "openai>=2.9.0",
    "polars>=1.35.2",
    "prefect>=3.6.5",
    "python-dotenv>=1.2.1",
    "y0 @ git+https://github.com/y0-causal-inference/y0.git",
]

[build-system]
requires = ["uv_build>=0.8.24,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=9.0.2",
    "pytest-instafail>=0.5.0",
    "pytest-timeout>=2.3.1",
    "pytest-xdist>=3.6.1",
    "ruff>=0.9.0",
    "streamlit>=1.52.2",
    "streamlit-agraph>=0.0.45",
    "watchdog>=6.0.0",
    "dspy>=3.0.4",
    "modal>=1.3.2",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]
fix = true

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (needed for Prefect)
    "B904",   # raise from err (too noisy)
    "B905",   # zip() without strict= (often unnecessary)
    "SIM108", # use ternary operator (not always clearer)
    "RUF001", # ambiguous unicode (intentional for display)
    "RUF022", # __all__ not sorted (comments prevent auto-sort)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "ARG001",  # unused function arguments (pytest fixtures)
    "ARG002",  # unused method arguments (pytest fixtures)
]

[tool.ruff.lint.isort]
known-first-party = ["causal_ssm_agent"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.pytest.ini_options]
addopts = "-n auto --tb=short -ra --instafail --timeout=300 --durations=10"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
